Split NoodleR-specific server routes into noodler.routes.ts
noodle.routes.ts mixed the core Twitter-style feed routes with the
OnlyFans-style NoodleR mechanic. Move the private-account create/
delete/retry-identity/simulate-fans routes, the NoodleR hub, and
subscribe/unsubscribe/unlock into a new noodler.routes.ts, registered
alongside noodleRoutes at the same "/noodle" prefix so client URLs are
unchanged. The underlying generation machinery (stage-identity/avatar
generation, linked-identity resolution, image-prompt sanitization)
stays in noodle.routes.ts and is imported by the new file, since it's
also used by the core refresh loop — genuinely shared, not NoodleR-
exclusive, so moving it would only relocate the entanglement rather
than remove it. Pure route relocation, no behavior change; verified
with a clean tsc/eslint pass and a manual server boot (no duplicate-
route registration error).
