+++ packages/server/src/routes/index.ts
@@ -56,6 +56,7 @@ import { promptOverridesRoutes } from "./prompt-overrides.routes.js";
 import { csrfDiagnosticsRoutes } from "./csrf-diagnostics.routes.js";
 import { professorMariWorkspaceRoutes } from "./professor-mari-workspace.routes.js";
 import { noodleRoutes } from "./noodle.routes.js";
+import { noodlerRoutes } from "./noodler.routes.js";
 
 export async function registerRoutes(app: FastifyInstance) {
   await app.register(chatsRoutes, { prefix: "/api/chats" });
@@ -111,6 +112,7 @@ export async function registerRoutes(app: FastifyInstance) {
   await app.register(csrfDiagnosticsRoutes, { prefix: "/api/csrf" });
   await app.register(professorMariWorkspaceRoutes, { prefix: "/api/professor-mari/workspace" });
   await app.register(noodleRoutes, { prefix: "/api/noodle" });
+  await app.register(noodlerRoutes, { prefix: "/api/noodle" });
   if (process.env.MARINARA_LITE !== "true" && process.env.MARINARA_LITE !== "1") {
     await app.register(sidecarRoutes, { prefix: "/api/sidecar" });
   }
